backend/elastic_indexer/requirements.txt
elasticsearch
google-generativeai
